requirements.txt
beautifulsoup4
lxml
psutil
yiri-mirai
numpy
apscheduler
uvicorn
